requirements.txt
requests
google-generativeai
python-telegram-bot
python-dotenv
supabase
feedparser
yfinance
pandas_datareader
vnstock
